Refactor revisionChooser components #265
