+++ visualization/app/codeCharta/ui/revisionChooser/revisionChooser.component.scss
