+++ visualization/app/codeCharta/ui/revisionChooser/revisionChooserFileDropDown.component.scss
